update main revisao

+++ revisao_pedido_produto/src/src/Distribuidor.java
@@ -45,7 +45,7 @@ public class Distribuidor {
         return preco;
     }
 
-    public void setPreco(float preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
